Rename file with capitailize because there are classes
