+++ src/models/QueryRangeGroupBy.js
