+++ test/models/Query.spec.js
@@ -1,7 +1,7 @@
 import chai from 'chai';
 
-import Query from '../../src/models/query';
-import QueryAggregate from '../../src/models/queryAggregate';
+import Query from '../../src/models/Query';
+import QueryAggregate from '../../src/models/QueryAggregate';
 
 
 describe('Query', function() {
